lambda_functions/src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "David Armendariz", email = "david.adrian.a.p@gmail.com" }]
dependencies = ["Flask==3.0.3", "werkzeug>=3.1.3", "boto3>=1.37.12", "python-dotenv>=1.0.1"]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false
